Move Infrastructure Helpers and Documentation filters to Common (partial)
Moves HealthCheckServiceExtensions, FileSystemAppSettingsFileProvider, and
the three Swagger/Profile documentation filters from the v2 and v3 projects
into EdFi.Ods.AdminApi.Common, updating consumers accordingly.

ConstantsHelper.cs is intentionally NOT moved: it reads
Assembly.GetExecutingAssembly().GetName().Version, and each app assembly's
version differs (v2's is stamped with the real release version by
build.ps1's SetAdminApiAssemblyInfo; v3's is static). This value is
user-visible via the public /information endpoint, which reports a
distinct Build value per AdminApiMode and is asserted per-mode in
ReadInformationTest. Moving it to Common would make GetExecutingAssembly()
resolve to Common.dll for both v2 and v3, collapsing the two into one
reported build version. Flagging for follow-up rather than silently
changing what version string ships.
