+++ Application/EdFi.Ods.AdminApi.Common/EdFi.Ods.AdminApi.Common.csproj
@@ -10,6 +10,7 @@
     <ItemGroup>
         <PackageReference Include="Asp.Versioning.Http" />
         <PackageReference Include="AspNetCore.HealthChecks.NpgSql" />
+        <PackageReference Include="AspNetCore.HealthChecks.SqlServer" />
         <PackageReference Include="EdFi.Suite3.Admin.DataAccess" />
         <PackageReference Include="EdFi.Suite3.Security.DataAccess" />
         <PackageReference Include="EFCore.NamingConventions" />
