+++ Application/EdFi.Ods.AdminApi.Common/Infrastructure/Helpers/FileSystemAppSettingsFileProvider.cs
@@ -3,7 +3,7 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-namespace EdFi.Ods.AdminApi.Infrastructure.Helpers;
+namespace EdFi.Ods.AdminApi.Common.Infrastructure.Helpers;
 
 public interface IAppSettingsFileProvider
 {
